# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API 文档示例",
        "title": "Gin Web API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/login": {
            "post": {
                "description": "用户登录获取token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "1001": {
                        "description": "参数校验失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "1002": {
                        "description": "用户名或密码错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "description": "注册新用户账号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户注册信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.UserRegister"
                        }
                    }
                ],
                "responses": {
                    "1001": {
                        "description": "参数校验失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "1002": {
                        "description": "注册失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/comment": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查询文章下的所有评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论管理"
                ],
                "summary": "查询评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "文章ID",
                        "name": "postId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "1001": {
                        "description": "参数校验失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "1002": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "为文章创建评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论管理"
                ],
                "summary": "创建评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "评论信息",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.CommnetByPostId"
                        }
                    }
                ],
                "responses": {
                    "1001": {
                        "description": "参数校验失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "1002": {
                        "description": "评论失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "200": {
                        "description": "评论成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/post": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据标题查询单个文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "查询单个文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"我的第一篇文章\"",
                        "description": "文章标题",
                        "name": "title",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "用户ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据用户ID和文章ID更新文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "更新文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "文章更新信息",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.UpPostByUserId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "更新失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建新文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "创建文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "文章信息",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.CreatePost"
                        }
                    }
                ],
                "responses": {
                    "1001": {
                        "description": "参数校验失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "1002": {
                        "description": "创建失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据用户ID和文章ID删除文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "文章删除信息",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.UpPostByUserId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/post/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据用户ID查询所有文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "查询用户文章列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "用户ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "web.CommnetByPostId": {
            "type": "object",
            "required": [
                "postId"
            ],
            "properties": {
                "content": {
                    "description": "评论内容",
                    "type": "string",
                    "example": "这是一条评论"
                },
                "postId": {
                    "description": "文章ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "web.CreatePost": {
            "type": "object",
            "required": [
                "content",
                "title",
                "userId"
            ],
            "properties": {
                "content": {
                    "description": "文章内容",
                    "type": "string",
                    "example": "这是文章内容"
                },
                "title": {
                    "description": "文章标题",
                    "type": "string",
                    "example": "我的第一篇文章"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "web.UpPostByUserId": {
            "type": "object",
            "required": [
                "postId",
                "userId"
            ],
            "properties": {
                "content": {
                    "description": "文章内容",
                    "type": "string",
                    "example": "更新后的内容"
                },
                "postId": {
                    "description": "文章ID",
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "description": "文章标题",
                    "type": "string",
                    "example": "更新后的标题"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "web.UserLogin": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "example": "testuser"
                }
            }
        },
        "web.UserRegister": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "example": "test@example.com"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "example": "testuser"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}